split cli out of __init__
